Refactor Standard Relayer handling
