+++ src/pages/Tx/Information/index.tsx
@@ -1,13 +1,7 @@
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useSearchParams } from "react-router-dom";
-import { ethers } from "ethers";
 import { useRecoilState } from "recoil";
-import { ChainId, isEVMChain, parseVaa } from "@certusone/wormhole-sdk";
-import {
-  DeliveryInstruction,
-  parseEVMExecutionInfoV1,
-} from "@certusone/wormhole-sdk/lib/cjs/relayer";
-import { ChainId as ApiChainId } from "src/api";
+import { ChainId, isEVMChain } from "@certusone/wormhole-sdk";
 import { useEnvironment } from "src/context/EnvironmentContext";
 import {
   CCTP_MANUAL_APP_ID,
@@ -23,25 +17,15 @@ import {
   txType,
   UNKNOWN_APP_ID,
 } from "src/consts";
-import { Alert, Loader } from "src/components/atoms";
+import { Alert } from "src/components/atoms";
 import { formatUnits, parseAddress, parseTx } from "src/utils/crypto";
 import { formatDate } from "src/utils/date";
 import { formatNumber } from "src/utils/number";
 import { getChainName, getExplorerLink } from "src/utils/wormhole";
 import analytics from "src/analytics";
-import {
-  DeliveryLifecycleRecord,
-  isRedelivery,
-  parseGenericRelayerVaa,
-  populateDeliveryLifecycleRecordByVaa,
-} from "src/utils/genericRelayerVaaUtils";
-import { tryGetAddressInfo, tryGetRedeemTxn } from "src/utils/cryptoToolkit";
+import { tryGetRedeemTxn } from "src/utils/cryptoToolkit";
 import { getPorticoInfo } from "src/utils/wh-portico-rpc";
-import {
-  addressesInfoState,
-  showSourceTokenUrlState,
-  showTargetTokenUrlState,
-} from "src/utils/recoilStates";
+import { showSourceTokenUrlState, showTargetTokenUrlState } from "src/utils/recoilStates";
 import { GetBlockData } from "src/api/search/types";
 import { GetOperationsOutput } from "src/api/guardian-network/types";
 import { getTokenInformation } from "src/utils/fetchWithRPCsFallthrough";
@@ -54,22 +38,20 @@ import RelayerOverview from "./Overview/RelayerOverview";
 import AdvancedView from "./AdvancedView";
 
 import "./styles.scss";
-import { ARKHAM_CHAIN_NAME } from "src/utils/arkham";
 
 interface Props {
   blockData: GetBlockData;
-  extraRawInfo: any;
-  setTxData: (x: any) => void;
   data: GetOperationsOutput;
+  extraRawInfo: any;
   isRPC: boolean;
+  setTxData: (x: any) => void;
 }
 
-const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props) => {
+const Information = ({ blockData, data, extraRawInfo, isRPC, setTxData }: Props) => {
   const [searchParams, setSearchParams] = useSearchParams();
 
   const [showSourceTokenUrl] = useRecoilState(showSourceTokenUrlState);
   const [showTargetTokenUrl] = useRecoilState(showTargetTokenUrlState);
-  const [addressesInfo, setAddressesInfo] = useRecoilState(addressesInfoState);
 
   const [showOverview, setShowOverviewState] = useState(searchParams.get("view") !== "advanced");
   const setShowOverview = (show: boolean) => {
@@ -320,7 +302,6 @@ const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props)
     tokenAmount,
     tokenInfo,
     totalGuardiansNeeded,
-    addressesInfo,
     VAAId,
   };
 
@@ -400,7 +381,7 @@ const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props)
       setFoundRedeem(true);
 
       setTimeout(() => {
-        setIsGenericRelayerTx(false);
+        // setIsGenericRelayerTx(false);
         setTxData(newData);
       }, 2000);
     } else {
@@ -416,164 +397,11 @@ const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props)
 
   const isLatestBlockHigherThanVaaEmitBlock = lastFinalizedBlock > currentBlock;
 
-  // Arkham for Standard Relayer txns
-  const [checkedArkham, setCheckedArkham] = useState(false);
-  const checkAddressesWithArkham = async (
-    deliveryParsedSenderAddress: string,
-    deliveryParsedTargetAddress: string,
-    targetChainId: number,
-    deliveryParsedRefundAddress: string,
-    refundChainId: number,
-    deliveryParsedSourceProviderAddress: string,
-    parsedEmitterAddress: string,
-  ) => {
-    if (checkedArkham) return;
-    const deliveryParsedSenderAddressInfo =
-      deliveryParsedSenderAddress && fromChain && ARKHAM_CHAIN_NAME[fromChain as ApiChainId]
-        ? await tryGetAddressInfo(currentNetwork, deliveryParsedSenderAddress)
-        : null;
-
-    const deliveryParsedTargetAddressInfo =
-      deliveryParsedTargetAddress && targetChainId && ARKHAM_CHAIN_NAME[targetChainId as ApiChainId]
-        ? await tryGetAddressInfo(currentNetwork, deliveryParsedTargetAddress)
-        : null;
-
-    const deliveryParsedRefundAddressInfo =
-      deliveryParsedRefundAddress && refundChainId && ARKHAM_CHAIN_NAME[refundChainId as ApiChainId]
-        ? await tryGetAddressInfo(currentNetwork, deliveryParsedRefundAddress)
-        : null;
-
-    const deliveryParsedSourceProviderAddressInfo =
-      deliveryParsedSourceProviderAddress &&
-      targetChainId &&
-      ARKHAM_CHAIN_NAME[targetChainId as ApiChainId]
-        ? await tryGetAddressInfo(currentNetwork, deliveryParsedSourceProviderAddress)
-        : null;
-
-    const parsedEmitterAddressInfo =
-      parsedEmitterAddress && fromChain && ARKHAM_CHAIN_NAME[fromChain as ApiChainId]
-        ? await tryGetAddressInfo(currentNetwork, parsedEmitterAddress)
-        : null;
-
-    if (!checkedArkham) {
-      setAddressesInfo({
-        ...addressesInfo,
-        [deliveryParsedTargetAddress.toLowerCase()]: deliveryParsedTargetAddressInfo,
-        [deliveryParsedRefundAddress.toLowerCase()]: deliveryParsedRefundAddressInfo,
-        [deliveryParsedSourceProviderAddress.toLowerCase()]:
-          deliveryParsedSourceProviderAddressInfo,
-        [parsedEmitterAddress.toLowerCase()]: parsedEmitterAddressInfo,
-        [deliveryParsedSenderAddress.toLowerCase()]: deliveryParsedSenderAddressInfo,
-      });
-      setCheckedArkham(true);
-    }
-  };
-
-  // --- Automatic Relayer Detection and handling ---
-  const [genericRelayerInfo, setGenericRelayerInfo] = useState<DeliveryLifecycleRecord>(null);
-  const [loadingRelayers, setLoadingRelayers] = useState(false);
-  const getRelayerInfo = useCallback(async () => {
-    setLoadingRelayers(true);
-
-    // TODO: handle generic relayer non-vaa txns without rpcs
-    if (!vaa || !vaa.raw) {
-      setLoadingRelayers(false);
-      setIsGenericRelayerTx(false);
-      console.log("automatic relayer tx without vaa yet");
-      return;
-    }
-
-    populateDeliveryLifecycleRecordByVaa(environment, vaa.raw)
-      .then((result: DeliveryLifecycleRecord) => {
-        analytics.track("txDetail", {
-          appIds: [GR_APP_ID].join(", "),
-          chain: getChainName({
-            chainId: (result?.sourceChainId as any)
-              ? (result.sourceChainId as any)
-              : standarizedProperties?.fromChain
-              ? standarizedProperties?.fromChain
-              : 0,
-            network: currentNetwork,
-          }),
-          toChain: getChainName({
-            chainId: result?.targetTransaction?.targetChainId
-              ? result.targetTransaction?.targetChainId
-              : standarizedProperties?.toChain
-              ? standarizedProperties?.toChain
-              : 0,
-            network: currentNetwork,
-          }),
-        });
-
-        setGenericRelayerInfo(result);
-        setLoadingRelayers(false);
-      })
-      .catch((e: any) => {
-        setLoadingRelayers(false);
-        console.error("automatic relayer tx errored:", e);
-        setIsGenericRelayerTx(false);
-      });
-  }, [
-    environment,
-    vaa,
-    standarizedProperties?.fromChain,
-    standarizedProperties?.toChain,
-    currentNetwork,
-  ]);
-
-  const targetContract = environment.chainInfos.find(
-    a => a.chainId === fromChain,
-  )?.relayerContractAddress;
-
-  const [isGenericRelayerTx, setIsGenericRelayerTx] = useState(null);
-
-  useEffect(() => {
-    if (targetContract || parsedEmitterAddress) {
-      const isGeneric = targetContract?.toUpperCase() === parsedEmitterAddress?.toUpperCase();
-
-      setIsGenericRelayerTx(isGeneric && !appIds?.includes(NTT_APP_ID));
-      if (isGeneric) {
-        console.log("isGenericRelayerTx!!!");
-        getRelayerInfo();
-      }
-    }
-  }, [targetContract, parsedEmitterAddress, getRelayerInfo, appIds]);
-  // --- x ---
-
-  useEffect(() => {
-    if (!loadingRelayers && genericRelayerInfo) {
-      if (
-        genericRelayerInfo?.targetTransaction?.targetTxHash &&
-        data &&
-        data.STATUS !== "COMPLETED"
-      ) {
-        setTxData({
-          ...data,
-          STATUS: "COMPLETED",
-          targetChain: {
-            ...data?.targetChain,
-            timestamp: genericRelayerInfo?.targetTransaction?.targetTxTimestamp * 1000,
-            transaction: {
-              ...data?.targetChain?.transaction,
-              txHash: genericRelayerInfo.targetTransaction.targetTxHash,
-            },
-          },
-        });
-
-        return;
-      }
-    }
-  }, [data, genericRelayerInfo, loadingRelayers, setTxData]);
+  const isGenericRelayerTx = data.relayerInfo !== null;
 
   const Content = () => {
-    if (isGenericRelayerTx === null) {
-      return <Loader />;
-    }
-
     if (isGenericRelayerTx || isRelayerNTT) {
-      if (loadingRelayers) return <Loader />;
-
-      const deliveryStatus = genericRelayerInfo?.DeliveryStatus;
+      const deliveryStatus = data.relayerInfo?.DeliveryStatus;
 
       if (isRelayerNTT && showOverview) {
         return (
@@ -587,7 +415,7 @@ const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props)
         );
       }
 
-      if (!genericRelayerInfo?.vaa) {
+      if (!data.relayerInfo?.vaa) {
         if (showOverview) {
           return <Overview {...overviewAndDetailProps} />;
         } else {
@@ -595,233 +423,21 @@ const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props)
             <AdvancedView
               overviewAndDetailProps={overviewAndDetailProps}
               extraRawInfo={extraRawInfo}
-              lifecycleRecord={genericRelayerInfo}
+              lifecycleRecord={data.relayerInfo}
               data={data}
             />
           );
         }
       }
 
-      const vaa = genericRelayerInfo.vaa;
-      const parsedVaa = parseVaa(vaa);
-      const sourceTxHash = genericRelayerInfo.sourceTxHash;
-
-      const resultLogRegex =
-        deliveryStatus?.data?.delivery?.execution?.detail.match(/Status: ([^\r\n]+)/);
-      const resultLog = resultLogRegex ? resultLogRegex?.[1] : null;
-
-      const gasUsed = Number(deliveryStatus?.data?.delivery?.execution?.gasUsed);
-      const targetTxTimestamp = genericRelayerInfo?.targetTransaction?.targetTxTimestamp;
-
-      const { emitterAddress, emitterChain, guardianSignatures } = parsedVaa || {};
-
-      const bufferEmitterAddress = Buffer.from(emitterAddress).toString("hex");
-      const parsedEmitterAddress = parseAddress({
-        value: bufferEmitterAddress,
-        chainId: emitterChain as ChainId,
-      });
-
-      const totalGuardiansNeeded = currentNetwork === "MAINNET" ? 13 : 1;
-      const guardianSignaturesCount = Array.isArray(guardianSignatures)
-        ? guardianSignatures?.length || 0
-        : 0;
-
-      const fromChain = emitterChain;
-
-      const instruction = parseGenericRelayerVaa(parsedVaa);
-      const deliveryInstruction = instruction as DeliveryInstruction | null;
-      const isDelivery = deliveryInstruction && !isRedelivery(deliveryInstruction);
-
-      const decodeExecution = deliveryInstruction.encodedExecutionInfo
-        ? parseEVMExecutionInfoV1(deliveryInstruction.encodedExecutionInfo, 0)[0]
-        : null;
-      const gasLimit = decodeExecution ? decodeExecution.gasLimit : null;
-
-      if (!deliveryInstruction?.targetAddress) {
-        return (
-          <div className="tx-information-errored-info">
-            This is either not an Automatic Relayer VAA or something&apos;s wrong with it
-          </div>
-        );
-      }
-
-      const trunkStringsDecimal = (num: string, decimals: number) => {
-        const [whole, fraction] = num.split(".");
-        if (!fraction) return whole;
-        return `${whole}.${fraction.slice(0, decimals)}`;
-      };
-
-      const maxRefund = deliveryStatus?.data?.delivery?.maxRefund
-        ? Number(
-            trunkStringsDecimal(
-              ethers.utils.formatUnits(
-                deliveryStatus?.data?.delivery?.maxRefund,
-                deliveryStatus?.data?.delivery?.targetChainDecimals || 18,
-              ),
-              3,
-            ),
-          )
-        : 0;
-
-      const deliveryParsedTargetAddress = parseAddress({
-        value: Buffer.from(deliveryInstruction?.targetAddress).toString("hex"),
-        chainId: deliveryInstruction?.targetChainId as ChainId,
-      });
-
-      const deliveryParsedRefundAddress = parseAddress({
-        value: Buffer.from(deliveryInstruction?.refundAddress).toString("hex"),
-        chainId: deliveryInstruction?.refundChainId as ChainId,
-      });
-
-      const deliveryParsedRefundProviderAddress = parseAddress({
-        value: Buffer.from(deliveryInstruction?.refundDeliveryProvider).toString("hex"),
-        chainId: deliveryInstruction?.refundChainId as ChainId,
-      });
-
-      const deliveryParsedSenderAddress = parseAddress({
-        value: Buffer.from(deliveryInstruction?.senderAddress).toString("hex"),
-        chainId: fromChain as ChainId,
-      });
-
-      const deliveryParsedSourceProviderAddress = parseAddress({
-        value: Buffer.from(deliveryInstruction?.sourceDeliveryProvider).toString("hex"),
-        chainId: fromChain as ChainId,
-      });
-
-      const maxRefundText = () => {
-        return `${maxRefund} ${
-          environment.chainInfos.find(chain => chain.chainId === deliveryInstruction.targetChainId)
-            .nativeCurrencyName
-        }`;
-      };
-
-      const gasUsedText = () => {
-        return isNaN(gasUsed) ? `${gasLimit}` : `${gasUsed}/${gasLimit}`;
-      };
-
-      const receiverValueText = () => {
-        const receiverValue = trunkStringsDecimal(
-          ethers.utils.formatUnits(
-            deliveryStatus?.data?.instructions?.requestedReceiverValue,
-            deliveryStatus?.data?.delivery?.targetChainDecimals || 18,
-          ),
-          3,
-        );
-
-        return `${receiverValue} ${
-          environment.chainInfos.find(chain => chain.chainId === deliveryInstruction.targetChainId)
-            .nativeCurrencyName
-        }`;
-      };
-
-      const budgetText = () => {
-        if (deliveryStatus?.data?.delivery?.budget) {
-          return `${trunkStringsDecimal(
-            ethers.utils.formatUnits(
-              deliveryStatus?.data?.delivery?.budget,
-              deliveryStatus?.data?.delivery?.targetChainDecimals || 18,
-            ),
-            3,
-          )} ${
-            environment.chainInfos.find(
-              chain => chain.chainId === deliveryInstruction.targetChainId,
-            ).nativeCurrencyName
-          }`;
-        }
-
-        return "N/A";
-      };
-
-      const refundText = () => {
-        const refundAmountRegex = deliveryStatus?.data?.delivery?.execution?.detail.match(
-          /Refund amount:\s*([0-9.]+)/,
-        );
-        const refundAmount = refundAmountRegex ? refundAmountRegex?.[1] : null;
-
-        if (refundAmount)
-          return `${refundAmount} ${
-            environment.chainInfos.find(
-              chain => chain.chainId === deliveryInstruction.targetChainId,
-            ).nativeCurrencyName
-          }`;
-
-        return "";
-      };
-
-      const copyBudgetText = () => {
-        return `Budget: ${budgetText()}\n\nMax Refund:\n${maxRefundText()}\n\n${
-          !isNaN(gasUsed) ? "Gas Used/" : ""
-        }Gas limit\n${gasUsedText()}\n\n${
-          !isNaN(gasUsed) ? "Refund Amount\n" + refundText() : ""
-        }\n\nReceiver Value: ${receiverValueText()}`
-          .replaceAll("  ", "")
-          .replaceAll("\n\n\n\n", "\n\n");
-      };
-
-      const refundStatusRegex =
-        deliveryStatus?.data?.delivery?.execution?.detail.match(/Refund status: ([^\r\n]+)/);
-      const refundStatus = refundStatusRegex ? refundStatusRegex?.[1] : null;
-
-      const deliveryAttemptRegex = deliveryStatus?.data?.delivery?.execution?.detail.match(
-        /Delivery attempt \s*([0-9.]+)/,
-      );
-      const deliveryAttempt = deliveryAttemptRegex ? deliveryAttemptRegex?.[1] : null;
-
-      const genericRelayerProps = {
-        addressesInfo,
-        budgetText,
-        copyBudgetText,
-        currentNetwork,
-        decodeExecution,
-        deliveryAttempt,
-        deliveryInstruction,
-        deliveryParsedRefundAddress,
-        deliveryParsedRefundProviderAddress,
-        deliveryParsedSenderAddress,
-        deliveryParsedSourceProviderAddress,
-        deliveryParsedTargetAddress,
-        deliveryStatus,
-        fromChain,
-        gasUsed,
-        gasUsedText,
-        guardianSignaturesCount,
-        isDelivery,
-        isDuplicated,
-        maxRefundText,
-        parsedEmitterAddress,
-        parsedVaa,
-        receiverValueText,
-        refundStatus,
-        refundText,
-        resultLog,
-        sourceTxHash,
-        targetTxTimestamp,
-        totalGuardiansNeeded,
-        VAAId,
-      };
-
-      checkAddressesWithArkham(
-        deliveryParsedSenderAddress,
-        deliveryParsedTargetAddress,
-        deliveryInstruction.targetChainId,
-        deliveryParsedRefundAddress,
-        deliveryInstruction.refundChainId,
-        deliveryParsedSourceProviderAddress,
-        parsedEmitterAddress,
-      );
-
       if (showOverview) {
-        return <RelayerOverview {...genericRelayerProps} />;
+        return <RelayerOverview {...data.relayerInfo.props} />;
       } else {
-        if (isGenericRelayerTx === null || (isGenericRelayerTx && loadingRelayers)) {
-          return <Loader />;
-        }
-
         return (
           <AdvancedView
-            genericRelayerProps={genericRelayerProps}
+            genericRelayerProps={data.relayerInfo.props}
             extraRawInfo={extraRawInfo}
-            lifecycleRecord={genericRelayerInfo}
+            lifecycleRecord={data.relayerInfo}
             data={data}
           />
         );
@@ -836,7 +452,7 @@ const Information = ({ blockData, extraRawInfo, setTxData, data, isRPC }: Props)
           <AdvancedView
             overviewAndDetailProps={overviewAndDetailProps}
             extraRawInfo={extraRawInfo}
-            lifecycleRecord={genericRelayerInfo}
+            lifecycleRecord={data.relayerInfo}
             data={data}
           />
         );
